Second (quick) pass on Free Lance
I've now skimmed quickly through Basil Langton's 1974 concert-stage
adaptation of "The Free Lance," in the form of a typescript from
the Sousa Archives at UIUC (thanks to Scott W. Schwartz of that
institution for scanning it for me). It has completely different
(and shorter) dialogue, which I haven't attempted to incorporate here.